Move to LongPressState.DragCanclled in dragInProgress action
