+++ src/hooks/useDragAndDropThought.tsx
@@ -9,7 +9,6 @@ import DragThoughtZone from '../@types/DragThoughtZone'
 import Path from '../@types/Path'
 import SimplePath from '../@types/SimplePath'
 import State from '../@types/State'
-import { longPress } from '../actions'
 import { addMulticursorActionCreator as addMulticursor } from '../actions/addMulticursor'
 import { alertActionCreator as alert } from '../actions/alert'
 import { createThoughtActionCreator as createThought } from '../actions/createThought'
@@ -17,6 +16,7 @@ import { dragHoldActionCreator as dragHold } from '../actions/dragHold'
 import { dragInProgressActionCreator as dragInProgress } from '../actions/dragInProgress'
 import { errorActionCreator as error } from '../actions/error'
 import { importFilesActionCreator as importFiles } from '../actions/importFiles'
+import { longPressActionCreator as longPress } from '../actions/longPress'
 import { moveThoughtActionCreator as moveThought } from '../actions/moveThought'
 import { setIsMulticursorExecutingActionCreator as setIsMulticursorExecuting } from '../actions/setIsMulticursorExecuting'
 import { ThoughtContainerProps } from '../components/Thought'
@@ -122,7 +122,7 @@ const endDrag = () => {
           dispatch(alert(null))
         }
       },
-      longPress({ value: LongPressState.DragCancelled }),
+      longPress({ value: LongPressState.Inactive }),
     ])
   })
 }
